requirements.txt
polarion~=1.4.0
requests~=2.32.3
python-dotenv~=1.0.1
python-certifi-win32
openai~=1.40.6
termcolor~=2.4.0
gradio~=4.43.0
pandas~=2.0.3
zeep~=4.2.1